Migrated Vector3 cpp implementation in math_utils.cpp
